+++ docs/changelog.rst
@@ -6,6 +6,26 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog`_, and this project adheres to
 `Semantic Versioning`_.
 
+Unreleased
+----------
+
+Changed
+~~~~~~~
+
+- :meth:`xugrid.UgridDataArrayAccessor.from_structured` previously required the
+  literal dimensions ``("y", "x")``. This requirement has been relaxed, it will
+  now infer the dimensions from the provided coordinates.
+- :meth:`xugrid.Ugrid2d.from_structured` previously only supported 1D
+  coordinates; it now detects whether coordinates are 1D or 2D automatically.
+  Accordingly, :meth:`xugrid.Ugrid2d.from_structured_multicoord` should no
+  longer be used, and calling it will give a FutureWarning.
+
+Added
+~~~~~
+
+- :meth:`xugrid.UgridDataset.from_structured` has been added to create
+  UgriDatasets from xarray Datasets.
+
 [0.12.1] 2024-09-09
 -------------------
 
@@ -78,7 +98,7 @@ Changed
 - Selection operations such as :meth:`UgridDataArrayAccessor.sel_points` will
   now also return points that are located on the edges of 2D topologies.
 - :attr:`xugrid.Ugrid1d.dimensions` and :attr:`xugrid.Ugrid2d.dimensions` now
-  raise a FutureWarning; use ``.dims`` or ``.sizes`` instead.
+  give a FutureWarning; use ``.dims`` or ``.sizes`` instead.
 - Improved performance of :func:`xugrid.open_dataset` and
   :func:`xugrid.merge_partitions` when handling datasets with a large number
   of variables (>100).